Day36/fullstack_dashboard/backend/requirements.txt
fastapi uvicorn sqlmodel
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
email-validator